Migrated TimerA1_ISR into moo_asm
Made a small bug fix to TimerA1_ASR and then integrated it into the
assembler file. Created a new .c file, build_asm.c, which holds the
original code that was compiled into asm, in case we need to generate it
again.

With this, all of the functions comprised mostly of inline assembly have
been migrated into the asm file.